compiles with -Wall, zmq subdir compiles with -pedantic
git-svn-id: http://svn.fastmq.org/svn/zeromq/branches/refactoring-for-0.3@279 7451a71c-b6c8-49cc-8254-141fc5361860

+++ examples/exchange/stat.cpp
@@ -61,8 +61,7 @@ public:
 
     inline void throughput (uint8_t meter_id, uint64_t throughput)
     {
-        printf ("%1d:%08llu\n",
-            (int) meter_id, throughput);
+        printf ("%1d:%08lld\n", (int) meter_id, (long long) throughput);
         fflush (stdout);
     }
 
@@ -71,8 +70,8 @@ public:
     {
         if (meter_id == 5) {
             if (!confirmation_timestamps.empty ()) {
-                printf ("l:%08llu\n",
-                    confirmation_timestamps.front () - timestamp);
+                printf ("l:%08lld\n",
+                    (long long) (confirmation_timestamps.front () - timestamp));
                 fflush (stdout);
                 confirmation_timestamps.pop ();
             }
@@ -83,8 +82,8 @@ public:
 
         if (meter_id == 6) {
             if (!order_timestamps.empty ()) {
-                printf ("l:%08llu\n",
-                    timestamp - order_timestamps.front ());
+                printf ("l:%08lld\n",
+                    ((long long) (timestamp - order_timestamps.front ())));
                 fflush (stdout);
                 order_timestamps.pop ();
             }
